fix:#15 handleResize was deleted, toEm isnt defined
